components moved to views for Refactoring purposes
