+++ __tests__/frontend/navbar.test.jsx
@@ -1,6 +1,6 @@
 import { test,expect } from 'vitest'
 import { render, screen } from '@testing-library/react'
-import NavBar from '../../app/components/Navbar'
+import NavBar from '../../app/views/Navbar'
 
 test('Settings link is present', () => {
     render(<NavBar />)
