+++ app/[courseCode]/assignments/page.jsx
@@ -1,14 +1,14 @@
 'use client'
-import CourseNavBar from '../../components/CourseNavBar';
-import Sidebar from '../../components/Sidebar';
+import CourseNavBar from '../../views/CourseNavBar';
+import Sidebar from '../../views/Sidebar';
 import { useEffect, useState } from 'react';
 import { onAuthStateChanged } from 'firebase/auth';
 import { auth } from '../../lib/firebase';
 import { getDoc, doc,getDocs,query,collection, where,updateDoc } from 'firebase/firestore';
 import db from '../../lib/firebase'
-import StudentAssignmentCard from '../../components/StudentAssignmentCard';
-import TeacherAssignmentCard from '../../components/TeacherAssignmentCard';
-import Loader from '../../components/Loader';
+import StudentAssignmentCard from '../../views/StudentAssignmentCard';
+import TeacherAssignmentCard from '../../views/TeacherAssignmentCard';
+import Loader from '../../views/Loader';
 
 import { createUser, getTeacherDoc } from '../../models/User';
 import { getCourseDoc } from '../../models/Course';
