+++ app/[courseCode]/grades/[name]/page.jsx
@@ -1,8 +1,8 @@
 'use client';
 import { useState, useEffect } from 'react';
-import CourseNavBar from '../../../components/CourseNavBar';
-import Sidebar from '../../../components/Sidebar';
-import Loader from '../../../components/Loader';
+import CourseNavBar from '../../../views/CourseNavBar';
+import Sidebar from '../../../views/Sidebar';
+import Loader from '../../../views/Loader';
 import { FaChevronDown } from 'react-icons/fa';
 import db from '../../../lib/firebase';
 import { Auth } from 'firebase/auth';
