+++ app/[courseCode]/quiz/page.jsx
@@ -1,9 +1,9 @@
 
 'use client'
 import { useState, useEffect } from 'react';
-import Sidebar from '../../components/Sidebar';
-import Loader from '../../components/Loader';
-import CourseNavBar from '../../components/CourseNavBar';
+import Sidebar from '../../views/Sidebar';
+import Loader from '../../views/Loader';
+import CourseNavBar from '../../views/CourseNavBar';
 import db from '../../lib/firebase';
 import {doc,setDoc} from 'firebase/firestore';
