+++ app/views/CourseCard.jsx
