+++ app/views/CourseHomeView.jsx
@@ -1,4 +1,4 @@
-import Loader from '../components/Loader';
+import Loader from './Loader';
 
 export default function CourseHomeView({ courseName = "Course Name", courseCode = "Course Code", loading, pdfUrl, userType = "Teacher"}) {
     return (
