+++ app/views/CourseNavBar.jsx
