+++ app/views/Navbar.jsx
