+++ app/views/StuCourseNavBar.jsx
@@ -1,4 +1,4 @@
-    'use client';
+'use client';
 import { useState } from 'react';
 import Link from 'next/link';
 import { FiMenu } from 'react-icons/fi';
