+++ app/views/StudentAssignmentCard.jsx
